rename ruleConfig to validatorConfig, move to validator file
